Fix organizer link in header, organize views according to different type of users, start fixing redirects when organizers edit study groups
