+++ studygroups/templatetags/extras.py
@@ -1,5 +1,7 @@
 from django import template
 
+from studygroups.models import TeamMembership
+
 import datetime
 
 register = template.Library()
@@ -14,4 +16,11 @@ def first_weekday_date(date):
     """
     week_start = date - datetime.timedelta(days=date.weekday())
     return week_start.date()
+
+@register.filter
+def is_organizer(user):
+    """
+    Return true if the user is an organizer of staff
+    """
+    return user.is_staff or TeamMembership.objects.filter(user=user, role=TeamMembership.ORGANIZER)
